# server/src/routes/misc.js
const express = require("express");
const router = express.Router();
const pool = require("../db/pool");
const { sendSubscriptionWelcomeEmail, sendReferralUpgradeEmail, sendMail } = require("../services/emailService");
const { verifyMonnifyWebhookSignature } = require("../utils/monnify");
const { restoreOrderStock } = require("../utils/orderStock");
const crypto = require("crypto");

function generateReferralCode() {
  return "BF-" + crypto.randomBytes(3).toString("hex").toUpperCase();
}

// ── CONTACT FORM ──────────────────────────────────────────────
router.post("/contact", async (req, res, next) => {
  const { name, email, phone, message } = req.body;

  if (!name?.trim() || !email?.trim() || !message?.trim()) {
    return res.status(400).json({ message: "Name, email and message are required" });
  }
  if (!email.includes("@") || !email.includes(".")) {
    return res.status(400).json({ message: "Please enter a valid email address" });
  }

  try {
    await pool.query(
      `INSERT INTO contact_messages (name, email, phone, message, created_at)
       VALUES ($1, $2, $3, $4, NOW())`,
      [name.trim(), email.trim().toLowerCase(), phone?.trim() || null, message.trim()]
    );

    const supportEmail = process.env.SUPPORT_EMAIL || "info@bemsfarms.com";
    sendMail({
      to: supportEmail,
      subject: `New contact message from ${name.trim()}`,
      html: `
        <p><strong>From:</strong> ${name.trim()} (${email.trim()})</p>
        ${phone ? `<p><strong>Phone:</strong> ${phone.trim()}</p>` : ""}
        <p><strong>Message:</strong></p>
        <p>${message.trim().replace(/\n/g, "<br>")}</p>
      `,
    }).catch((err) => console.error("⚠️ Contact notification email failed:", err.message));

    res.json({ success: true, message: "Message sent! We'll get back to you within 24 hours." });
  } catch (err) {
    next(err);
  }
});

// ── SUBSCRIBE ─────────────────────────────────────────────────
router.post("/subscribe", async (req, res, next) => {
  const { email, referred_by } = req.body;

  // Validate
  if (!email) {
    return res.status(400).json({ message: "Email is required" });
  }
  if (
    typeof email !== "string" ||
    !email.includes("@") ||
    !email.includes(".")
  ) {
    return res
      .status(400)
      .json({ message: "Please enter a valid email address" });
  }

  const cleanEmail = email.toLowerCase().trim();
  const cleanReferredBy = referred_by ? referred_by.trim() : null;

  try {
    // 1. Check if email already exists
    const checkRes = await pool.query(
      "SELECT id, email, referral_code, referral_count, discount_code FROM email_subscriptions WHERE email = $1",
      [cleanEmail]
    );

    if (checkRes.rows.length > 0) {
      const existing = checkRes.rows[0];
      return res.json({
        success: true,
        message: `Welcome back! You are already subscribed. Your discount code is ${existing.discount_code}.`,
        code: existing.discount_code,
        referral_code: existing.referral_code,
        referral_count: existing.referral_count,
      });
    }

    // 2. Generate a unique referral code
    let referralCode = generateReferralCode();
    let isUnique = false;
    let attempts = 0;

    // Safety loop to ensure uniqueness
    while (!isUnique && attempts < 5) {
      const codeCheck = await pool.query(
        "SELECT id FROM email_subscriptions WHERE referral_code = $1",
        [referralCode]
      );
      if (codeCheck.rows.length === 0) {
        isUnique = true;
      } else {
        referralCode = generateReferralCode();
        attempts++;
      }
    }

    // 3. Check if referred by someone valid
    let finalReferredBy = null;
    if (cleanReferredBy) {
      const referrerCheck = await pool.query(
        "SELECT referral_code FROM email_subscriptions WHERE referral_code = $1",
        [cleanReferredBy]
      );
      if (referrerCheck.rows.length > 0) {
        finalReferredBy = referrerCheck.rows[0].referral_code;
      }
    }

    // 4. Insert new subscription
    await pool.query(
      `INSERT INTO email_subscriptions (email, referral_code, referred_by, discount_code, is_active, subscribed_at)
       VALUES ($1, $2, $3, 'BEMS10', true, NOW())`,
      [cleanEmail, referralCode, finalReferredBy]
    );

    // 5. Update referrer count & check threshold rewards
    if (finalReferredBy) {
      try {
        const referrerUpdate = await pool.query(
          `UPDATE email_subscriptions 
           SET referral_count = referral_count + 1 
           WHERE referral_code = $1 
           RETURNING email, referral_count`,
          [finalReferredBy]
        );

        if (referrerUpdate.rows.length > 0) {
          const referrer = referrerUpdate.rows[0];

          // Trigger rewards upgrade — >= thresholds (not ===) so the
          // referrer still gets upgraded if the count ever advances by more
          // than 1 between checks; the discount_code comparison in the
          // UPDATE below keeps this idempotent so it doesn't re-fire (and
          // re-email) on every signup past the threshold.
          let upgradeCode = null;
          if (referrer.referral_count >= 5) {
            upgradeCode = "BEMS30";
          } else if (referrer.referral_count >= 3) {
            upgradeCode = "BEMS20";
          }

          if (upgradeCode) {
            const upgraded = await pool.query(
              "UPDATE email_subscriptions SET discount_code = $1 WHERE referral_code = $2 AND discount_code IS DISTINCT FROM $1 RETURNING email",
              [upgradeCode, finalReferredBy]
            );
            if (upgraded.rows.length > 0) {
              await sendReferralUpgradeEmail(referrer.email, referrer.referral_count, upgradeCode);
            }
          }
        }
      } catch (refErr) {
        console.warn("⚠️ Referrer update failed:", refErr.message);
      }
    }

    // 6. Send welcome email to joiner
    try {
      await sendSubscriptionWelcomeEmail(cleanEmail, referralCode, "BEMS10");
    } catch (emailErr) {
      console.warn("⚠️ Welcome email sending failed:", emailErr.message);
    }

    res.json({
      success: true,
      message: "Subscribed! Use code BEMS10 for 10% off your next order.",
      code: "BEMS10",
      referral_code: referralCode,
      referral_count: 0
    });

  } catch (err) {
    console.error("❌ Subscribe DB error:", err.message);
    next(err);
  }
});

// ── MONNIFY WEBHOOK ──────────────────────────────────────────
// Built against Monnify's documented webhook shape
// (https://developers.monnify.com/docs/integration-tools/webhooks/) — not
// yet exercised against a real sandbox event. Verify the exact eventType
// values and eventData field names against a live test webhook once real
// MONNIFY_SECRET_KEY is configured, the same way the Paystack webhook was
// verified end-to-end earlier.
router.post(
  "/webhooks/monnify",
  async (req, res, next) => {
    console.log("🔔 Monnify webhook received");
    const signature = req.headers["monnify-signature"];

    if (!process.env.MONNIFY_SECRET_KEY && process.env.NODE_ENV === "production") {
      console.error("❌ MONNIFY_SECRET_KEY is not configured — refusing webhook (fail closed)");
      return res.status(503).json({ message: "Webhook processing not configured" });
    }

    let event;
    let rawPayload;

    try {
      rawPayload = req.rawBody || Buffer.from(typeof req.body === "string" ? req.body : JSON.stringify(req.body));
      event = typeof req.body === "string" ? JSON.parse(req.body) : req.body;
      if (req.rawBody) {
        event = JSON.parse(req.rawBody.toString("utf8"));
      }
    } catch (parseErr) {
      console.error("❌ Failed to parse Monnify webhook body:", parseErr.message);
      return res.status(400).json({ message: "Invalid JSON body" });
    }

    const isVerified = verifyMonnifyWebhookSignature(rawPayload, signature);
    // transactionReference is what we store as orders.payment_ref (see
    // orders.js / CheckoutPage.jsx); paymentReference is our own
    // client-generated tracking reference, kept only for the audit log.
    const reference = event?.eventData?.transactionReference || null;
    const eventType = event?.eventType || null;

    if (!isVerified) {
      console.warn("⚠️ Webhook signature mismatch!");
      // Save unverified audit log for safety
      try {
        await pool.query(
          `INSERT INTO payment_webhook_logs (event_type, payment_ref, payload, signature_verified, status, error_message)
           VALUES ($1, $2, $3, false, 'error', 'Signature verification failed')`,
          [eventType, reference, event]
        );
      } catch (dbErr) {
        console.error("Failed to write audit log:", dbErr.message);
      }
      return res.status(401).json({ message: "Signature verification failed" });
    }

    try {
      // 1. Idempotency Check: Do we already have this payment reconciled?
      const existingPay = await pool.query(
        "SELECT id, status FROM payments WHERE payment_ref = $1",
        [reference]
      );

      if (existingPay.rows.length > 0) {
        const payRecord = existingPay.rows[0];
        console.log(`ℹ️ Payment reference ${reference} already exists in DB with status: ${payRecord.status}`);
        
        // Log the duplicate event
        await pool.query(
          `INSERT INTO payment_webhook_logs (event_type, payment_ref, payload, signature_verified, status, error_message)
           VALUES ($1, $2, $3, true, 'ignored', 'Duplicate event received')`,
          [eventType, reference, event]
        );
        
        return res.status(200).json({ message: "Duplicate payment reference ignored" });
      }

      // 2. Process Successful, Failed, or Reversed transaction
      const eventData = event?.eventData || {};
      let status = "pending";
      if (eventType === "SUCCESSFUL_TRANSACTION" || eventData.paymentStatus === "PAID") {
        status = "successful";
      } else if (eventType === "FAILED_TRANSACTION" || eventData.paymentStatus === "FAILED") {
        status = "failed";
      } else if (eventType?.toUpperCase().includes("REVERS") || eventType?.toUpperCase().includes("REFUND")) {
        status = "reversed";
      }

      // Monnify amounts are plain Naira decimals, not kobo.
      const amount = parseFloat(eventData.amountPaid || 0);
      const email = eventData.customer?.email || null;
      const metadata = eventData.metaData || eventData.metadata || null;
      const paidAt = eventData.paidOn ? new Date(eventData.paidOn) : null;
      const paymentMethod = eventData.paymentMethod || null;
      const terminalId = metadata?.pos_terminal_id || null;

      // 3. Find matching order in Bems Farms database
      const orderSearch = await pool.query(
        "SELECT id, total FROM orders WHERE payment_ref = $1 OR id::text = $2",
        [reference, String(metadata?.order_id || "")]
      );
      
      const linkedOrder = orderSearch.rows[0];
      const orderId = linkedOrder ? linkedOrder.id : null;

      // 4. Save validated payment details in DB
      await pool.query(
        `INSERT INTO payments (payment_ref, order_id, amount, status, payment_method, customer_email, pos_terminal_id, metadata, paid_at, updated_at)
         VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, NOW())
         ON CONFLICT (payment_ref) 
         DO UPDATE SET status = EXCLUDED.status, updated_at = NOW()`,
        [reference, orderId, amount, status, paymentMethod, email, terminalId, event, paidAt]
      );

      // 5. Automatic Reconciliation actions
      if (status === "successful") {
        if (orderId) {
          const orderTotal = parseFloat(linkedOrder.total) || 0;
          if (Math.abs(amount - orderTotal) > 0.01) {
            console.warn(
              `⚠️ Amount mismatch for order ${orderId}: paid ₦${amount} vs order total ₦${orderTotal}. Not auto-confirming — flagged for manual review.`
            );
            await pool.query(
              `INSERT INTO payment_webhook_logs (event_type, payment_ref, payload, signature_verified, status, error_message)
               VALUES ($1, $2, $3, true, 'error', $4)`,
              [eventType, reference, event, `Amount mismatch: paid ${amount} vs order total ${orderTotal}`]
            );
            return res.status(200).json({
              success: false,
              message: "Amount mismatch — order not auto-confirmed, flagged for manual review",
            });
          }

          // Reconcile order as 'confirmed'
          await pool.query(
            "UPDATE orders SET status = 'confirmed', payment_method = $1, payment_ref = $2, updated_at = NOW() WHERE id = $3",
            [paymentMethod || "transfer", reference, orderId]
          );
          console.log(`✅ Order ${orderId} successfully reconciled automatically with payment ref ${reference}`);

          // Fetch system user for ledger attribution
          let systemUserId = null;
          const systemUserRes = await pool.query(
            "SELECT id FROM users ORDER BY (CASE WHEN role='superadmin' THEN 1 WHEN role='manager' THEN 2 WHEN role='admin' THEN 3 ELSE 4 END) LIMIT 1"
          );
          if (systemUserRes.rows.length > 0) {
            systemUserId = systemUserRes.rows[0].id;
          }

          // Register in general accounting ledger (income table)
          await pool.query(
            `INSERT INTO income (reference, source, source_type, category, description, amount, payment_method, order_id, status, date, created_by)
             VALUES ($1, 'sales', 'online_order', 'POS/Online Sale', $2, $3, $4, $5, 'completed', CURRENT_DATE, $6)
             ON CONFLICT (reference) DO NOTHING`,
            [`INC-${reference}`, `Automated payment reconciliation for Order #${orderId}`, amount, paymentMethod || "transfer", String(orderId), systemUserId]
          );
        } else {
          console.log(`⚠️ Payment ref ${reference} received but no matching order ID was found in DB metadata.`);
        }
      } else if (status === "reversed") {
        if (orderId) {
          const prevOrder = await pool.query("SELECT status FROM orders WHERE id = $1", [orderId]);
          const alreadyCancelled = prevOrder.rows[0]?.status === "cancelled";

          // Update order to cancelled/refunded
          await pool.query(
            "UPDATE orders SET status = 'cancelled', updated_at = NOW() WHERE id = $1",
            [orderId]
          );
          // Reverse corresponding ledger income entry
          await pool.query(
            "UPDATE income SET status = 'reversed' WHERE order_id = $1",
            [String(orderId)]
          );
          // Restore stock deducted at order creation — same as every other
          // order-cancellation path. Guard against a duplicate webhook
          // delivery double-restoring stock for an order already cancelled.
          if (!alreadyCancelled) {
            await restoreOrderStock(pool, orderId);
          }
          console.log(`🔄 Order ${orderId} reversed/cancelled following webhook refund notification.`);
        }
      }

      // 6. Audit Log: Log processed webhook event
      await pool.query(
        `INSERT INTO payment_webhook_logs (event_type, payment_ref, payload, signature_verified, status)
         VALUES ($1, $2, $3, true, 'processed')`,
        [eventType, reference, event]
      );

      res.status(200).json({ success: true, message: "Webhook processed successfully" });
    } catch (err) {
      console.error("❌ Webhook db processor failed:", err.message);
      // Log the crash audit event
      try {
        await pool.query(
          `INSERT INTO payment_webhook_logs (event_type, payment_ref, payload, signature_verified, status, error_message)
           VALUES ($1, $2, $3, $4, 'error', $5)`,
          [eventType, reference, event, true, err.message]
        );
      } catch (dbErr) {
        console.error("Failed to write audit log:", dbErr.message);
      }
      next(err);
    }
  }
);

module.exports = router;